# openapi.yml
openapi: 3.0.1
info:
  title: Seltorn
  description: This is Seltor's chat API documentation. You will need a Bearer Token for most requests as the Authorization header. You can get it by the login request.
  version: 0.0.0
servers:
  - url: http://localhost:3000
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
...
paths:
  /login:
    post:
      tags:
      - Accounts
      summary: Logs in to your account
      requestBody:
        description: OTP is not always needed
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
                otp:
                  type: string
        required: true
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        401:
          description: Invalid information
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Invalid information."
        403:
          description: Account not verified
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Account not verified."
  /register:
    post:
      tags:
      - Accounts
      summary: Creates a new account
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                username:
                  type: string
                password:
                  type: string
        required: true
      responses:
        201:
          description: Success
          content: {}
        400:
          description: Request body's missing something
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Something is missing or it's not appropiate."
        401:
          description: Email in use by a verified account
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Email in use."
...